fix: tests

+++ tests/setup.ts
@@ -1,4 +1,5 @@
 import * as path from 'path';
+import * as os from 'os';
 import * as fs from 'fs-extra';
 import { beforeEach, afterEach } from '@jest/globals';
 
@@ -7,17 +8,18 @@ declare global {
   var testDir: string;
 }
 
-// Create temp directory for each test
+// Create unique temp directory for each test
 beforeEach(() => {
-  const tmpDir = path.join(__dirname, '..', '.tmp-test');
+  const uniqueDir = `.tmp-test-${Date.now()}-${Math.random().toString(36).substring(7)}`;
+  const tmpDir = path.join(os.tmpdir(), uniqueDir);
   fs.ensureDirSync(tmpDir);
   global.testDir = tmpDir;
 });
 
 // Cleanup temp directory after each test
 afterEach(() => {
-  const tmpDir = path.join(__dirname, '..', '.tmp-test');
-  if (fs.existsSync(tmpDir)) {
+  const tmpDir = global.testDir;
+  if (tmpDir && fs.existsSync(tmpDir)) {
     try {
       fs.removeSync(tmpDir);
     } catch (e) {
